Feat :Add Global Variable - 전역변수 추가했습니다. - 추가적으로 채팅 구현중입니다
